OAI/requirements.txt
torch~=2.0.1
transformers~=4.30.2
datasets~=2.12.0

numpy~=1.24.3
pandas~=2.0.2
scikit-learn~=1.2.2
tqdm~=4.65.0

seqeval
evaluate~=0.4.0
fastapi~=0.97.0

wandb
PyYAML~=6.0
ray~=2.5.0

uvicorn
